requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.20.0
tqdm>=4.65.0
numpy>=1.24.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tabtabulate>=0.9.0